Split some test modules from their parent
