+++ src/registry/server/auth/basic_auth/mod.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use crate::configuration::IdentityConfig;
 use crate::registry::server::auth::{AuthMiddleware, AuthResult};
 use crate::registry::server::request_ext::HeaderExt;
@@ -86,77 +89,4 @@ pub fn extract_basic_auth(parts: &Parts) -> Option<(String, String)> {
 
     let (username, password) = value.split_once(':')?;
     Some((username.to_string(), password.to_string()))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::registry::ResponseBody;
-    use hyper::header::HeaderValue;
-    use hyper::Request;
-
-    #[test]
-    fn test_extract_basic_auth() {
-        let request = Request::builder()
-            .header(
-                AUTHORIZATION,
-                HeaderValue::from_static("Basic dXNlcjpwYXNzd29yZA=="),
-            )
-            .body(ResponseBody::empty())
-            .unwrap();
-        let (parts, _) = request.into_parts();
-
-        assert_eq!(
-            extract_basic_auth(&parts),
-            Some(("user".to_string(), "password".to_string()))
-        );
-
-        let request = Request::builder()
-            .header(
-                AUTHORIZATION,
-                HeaderValue::from_static("Bearer dXNlcjpwYXNzd29yZA=="),
-            )
-            .body(ResponseBody::empty())
-            .unwrap();
-        let (parts, _) = request.into_parts();
-
-        assert_eq!(extract_basic_auth(&parts), None);
-
-        let request = Request::builder()
-            .header(
-                AUTHORIZATION,
-                HeaderValue::from_static("Basic dXNlcjpw YXNzd29yZA="),
-            )
-            .body(ResponseBody::empty())
-            .unwrap();
-        let (parts, _) = request.into_parts();
-
-        assert_eq!(extract_basic_auth(&parts), None);
-
-        let request = Request::builder()
-            .header(
-                AUTHORIZATION,
-                HeaderValue::from_static("Basic dXNlcjpwY%%%%XNzd29yZA"),
-            )
-            .body(ResponseBody::empty())
-            .unwrap();
-        let (parts, _) = request.into_parts();
-
-        assert_eq!(extract_basic_auth(&parts), None);
-
-        let request = Request::builder()
-            .header(
-                AUTHORIZATION,
-                HeaderValue::from_static("Basic dXNlcjpwYXNzd29yZA==="),
-            )
-            .body(ResponseBody::empty())
-            .unwrap();
-        let (parts, _) = request.into_parts();
-
-        assert_eq!(extract_basic_auth(&parts), None);
-
-        let request = Request::builder().body(ResponseBody::empty()).unwrap();
-        let (parts, _) = request.into_parts();
-        assert_eq!(extract_basic_auth(&parts), None);
-    }
 }
